Final Commit after completing code for both Tasks 3 and 5 by Jakob Soukthavong
